requirements/prod.txt
# dependencies for production
-r common.txt
supervisor==3.3.4
gunicorn==19.7.1
boto3==1.9.43
django-storages==1.7.1